CB-5574. Fixed update secret management, throw exception if secret management turned off
