+++ server/bundles/io.cloudbeaver.server/src/io/cloudbeaver/service/core/impl/WebServiceCore.java
@@ -342,109 +342,8 @@ public class WebServiceCore implements DBWServiceCore {
         boolean sharedCredentials,
         @Nullable String selectedSecretId
     ) throws DBWebException {
-        WebConnectionInfo connectionInfo = WebDataSourceUtils.getWebConnectionInfo(webSession, projectId, connectionId);
-        connectionInfo.setSavedCredentials(authProperties, networkCredentials);
-
-        var dataSourceContainer = (DataSourceDescriptor) connectionInfo.getDataSourceContainer();
-        if (dataSourceContainer.isConnected()) {
-            throw new DBWebException("Datasource '" + dataSourceContainer.getName() + "' is already connected");
-        }
-        if (dataSourceContainer.isSharedCredentials() && selectedSecretId != null) {
-            List<DBSSecretValue> allSecrets;
-            try {
-                allSecrets = dataSourceContainer.listSharedCredentials();
-            } catch (DBException e) {
-                throw new DBWebException("Error loading connection secret", e);
-            }
-            DBSSecretValue selectedSecret =
-                allSecrets.stream()
-                    .filter(secret -> selectedSecretId.equals(secret.getUniqueId()))
-                    .findFirst().orElse(null);
-            if (selectedSecret == null) {
-                throw new DBWebException("Secret not found:" + selectedSecretId);
-            }
-            dataSourceContainer.setSelectedSharedCredentials(selectedSecret);
-        }
-
-        boolean oldSavePassword = dataSourceContainer.isSavePassword();
-        DBRProgressMonitor monitor = webSession.getProgressMonitor();
-        validateDriverLibrariesPresence(dataSourceContainer);
-        try {
-            boolean connect = dataSourceContainer.connect(monitor, true, false);
-            if (connect) {
-                webSession.addSessionEvent(
-                    new WSDataSourceConnectEvent(
-                        projectId,
-                        connectionId,
-                        webSession.getSessionId(),
-                        webSession.getUserId()
-                    )
-                );
-            }
-        } catch (Exception e) {
-            throw new DBWebException("Error connecting to database", e);
-        } finally {
-            dataSourceContainer.setSavePassword(oldSavePassword);
-            connectionInfo.clearCache();
-        }
-        // Mark all specified network configs as saved
-        boolean[] saveConfig = new boolean[1];
-
-        if (networkCredentials != null) {
-            networkCredentials.forEach(c -> {
-                if (CommonUtils.toBoolean(c.isSavePassword())) {
-                    DBWHandlerConfiguration handlerCfg = dataSourceContainer.getConnectionConfiguration()
-                        .getHandler(c.getId());
-                    if (handlerCfg != null &&
-                        // check username param only for ssh config
-                        !(CommonUtils.isEmpty(c.getUserName()) && CommonUtils.equalObjects(handlerCfg.getType(),
-                            DBWHandlerType.TUNNEL))
-                    ) {
-                        WebDataSourceUtils.updateHandlerCredentials(handlerCfg, c);
-                        handlerCfg.setSavePassword(true);
-                        saveConfig[0] = true;
-                    }
-                }
-            });
-        }
-        if (saveCredentials) {
-            // Save all passed credentials in the datasource container
-            WebServiceUtils.saveAuthProperties(
-                dataSourceContainer,
-                dataSourceContainer.getConnectionConfiguration(),
-                authProperties,
-                true,
-                sharedCredentials
-            );
-
-            var project = dataSourceContainer.getProject();
-            if (project.isUseSecretStorage()) {
-                try {
-                    dataSourceContainer.persistSecrets(
-                        DBSSecretController.getProjectSecretController(dataSourceContainer.getProject())
-                    );
-                } catch (DBException e) {
-                    throw new DBWebException("Failed to save credentials", e);
-                }
-            }
-
-            WebDataSourceUtils.saveCredentialsInDataSource(connectionInfo,
-                dataSourceContainer,
-                dataSourceContainer.getConnectionConfiguration());
-            saveConfig[0] = true;
-        }
-        if (WebServiceUtils.isGlobalProject(dataSourceContainer.getProject())) {
-            // Do not flush config for global project (only admin can do it - CB-2415)
-            if (saveCredentials) {
-                connectionInfo.setCredentialsSavedInSession(true);
-            }
-            saveConfig[0] = false;
-        }
-        if (saveConfig[0]) {
-            dataSourceContainer.persistConfiguration();
-        }
-
-        return connectionInfo;
+        return WebAppUtils.getWebApplication().getConnectionController().initConnection(webSession, projectId,
+            connectionId, authProperties, networkCredentials, saveCredentials, sharedCredentials, selectedSecretId);
     }
 
     @Override
